İlk commit

+++ src/main/java/az/edu/itbrains/course/dtos/create/CourseCreateDto.java
@@ -20,5 +20,6 @@ public class CourseCreateDto {
 
     private boolean isPopular;
 
-    private int durationInHours;
+    private Double durationInHours;
+
 }
